Restore language configuration; default to English
The previous version removed the language configuration and just used
the system default, but some users prefer choosing the language.  This
restores the ability to choose, but keeps the system default as the
default (and, unlike before, allows "system default" as a choice).

This also changes the language default to "en" if the local language is
not found; the previous update accidentally changed it to "de" (because
"de" was the first in the list of languages, likely because it was
sorted alphabetically in the generated language code).

There is still a slight issue on the Settings page that isn't redrawn
instantly after a language change, so if you change language -> go Back
to settings, it will not have changed, but if you go back again and into
settings again everything will be updated, which seems good enough.