Add pending rewards and revamp balance display
This changes the balance display to show three balances:

- Full
- Unlocked
- Pending Rewards (with the next payment block beneath it; will not be
  shown if pending balance is 0).

Full can optionally include the batched pending rewards (in addition to
your current owned wallet outputs).

Gone is the tap-to-switch mode that was present before; it was a bit
unintuitive, and having to tap your finger on top of the value you are
trying to view made the UX pretty bad.

This replaces the single "Balance Display" setting (which was full,
available, or hidden) with 4 new independent options:
- show full
- show available
- show pending
- include pending in full

so that you can customize it to show as little or as much as you like.